abi: Move abi ahead of refactor
