+++ abi/bind/auth.go
