+++ signer/fourbyte/abi_test.go
@@ -22,7 +22,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/ethereum/go-ethereum/accounts/abi"
+	"github.com/ethereum/go-ethereum/abi"
 	"github.com/ethereum/go-ethereum/common"
 )
